Remove dependence on global PIXI mount point
Adds a new file which exposes registerPixiJs which doesn't assume that
PIXI is available as a global variable. For backwards compatibility Widget.js
does assume that PIXI is a global variable, but instead of shimming it as a
empty function, we raise an explicit error message. Trying to inherit from
the function raises an error anyways, but a less explicit one.

The original error reads:

`TypeError: Class extends value () => {} is not a constructor or null`

which requires inspecting the source code to discern the problem. With this
patch, it reads:

`Error: Cannot find PIXI`

which I think is much better.